requirements.txt
flask==3.0.0
requests==2.31.0
pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
newspaper4k==0.9.3
lxml_html_clean==0.1.1
huggingface_hub>=0.19.0